+++ contracts/lib/MMR.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.7;
 
-
 /**
  * @author Wanseob Lim <email@wanseob.com>
  * @title Merkle Mountain Range solidity library
@@ -10,7 +9,6 @@ pragma solidity ^0.7;
  *      And it uses keccak256 for its hash function instead of blake2b
  */
 library MMR {
-
     struct Tree {
         bytes32 root;
         uint256 size;
@@ -26,7 +24,7 @@ library MMR {
     function append(Tree storage tree, bytes memory data) public {
         // Hash the leaf node first
         bytes32 dataHash = keccak256(data);
-        if(keccak256(tree.data[dataHash]) != dataHash) {
+        if (keccak256(tree.data[dataHash]) != dataHash) {
             tree.data[dataHash] = data;
         }
         bytes32 leaf = hashLeaf(tree.size + 1, dataHash);
@@ -39,14 +37,18 @@ library MMR {
         tree.size = getSize(tree.width);
         // Starting from the left-most peak, get all peak hashes using _getOrCreateNode() function.
         bytes32[] memory peaks = new bytes32[](peakIndexes.length);
-        for (uint i = 0; i < peakIndexes.length; i++) {
+        for (uint256 i = 0; i < peakIndexes.length; i++) {
             peaks[i] = _getOrCreateNode(tree, peakIndexes[i]);
         }
         // Create the root hash and update the tree
         tree.root = peakBagging(tree.width, peaks);
     }
 
-    function hashOf(Tree storage tree, uint256 index) public view returns (bytes32) {
+    function hashOf(Tree storage tree, uint256 index)
+        public
+        view
+        returns (bytes32)
+    {
         return tree.hashes[index];
     }
 
@@ -64,7 +66,7 @@ library MMR {
         tree.size = getSize(tree.width);
         // Starting from the left-most peak, get all peak hashes using _getOrCreateNode() function.
         bytes32[] memory peaks = new bytes32[](peakIndexes.length);
-        for (uint i = 0; i < peakIndexes.length; i++) {
+        for (uint256 i = 0; i < peakIndexes.length; i++) {
             peaks[i] = _getOrCreateNode(tree, peakIndexes[i]);
         }
         // Create the root hash and update the tree
@@ -88,33 +90,37 @@ library MMR {
         tree.size = getSize(tree.width);
         // Starting from the left-most peak, get all peak hashes using _getOrCreateNode() function.
         bytes32[] memory peaks = new bytes32[](peakIndexes.length);
-        for (uint i = 0; i < peakIndexes.length; i++) {
+        for (uint256 i = 0; i < peakIndexes.length; i++) {
             peaks[i] = _getOrCreateNode(tree, peakIndexes[i]);
         }
         // Create the root hash and update the tree
         tree.root = peakBagging(tree.width, peaks);
     }
 
-    function getPeaks(Tree storage tree) public view returns (bytes32[] memory peaks) {
+    function getPeaks(Tree storage tree)
+        public
+        view
+        returns (bytes32[] memory peaks)
+    {
         // Find peaks for the enlarged tree
         uint256[] memory peakNodeIndexes = getPeakIndexes(tree.width);
         // Starting from the left-most peak, get all peak hashes using _getOrCreateNode() function.
         peaks = new bytes32[](peakNodeIndexes.length);
-        for (uint i = 0; i < peakNodeIndexes.length; i++) {
+        for (uint256 i = 0; i < peakNodeIndexes.length; i++) {
             peaks[i] = tree.hashes[peakNodeIndexes[i]];
         }
         return peaks;
     }
 
-    function getLeafIndex(uint width) public pure returns (uint) {
-        if(width % 2 == 1) {
+    function getLeafIndex(uint256 width) public pure returns (uint256) {
+        if (width % 2 == 1) {
             return getSize(width);
         } else {
             return getSize(width - 1) + 1;
         }
     }
 
-    function getSize(uint width) public pure returns (uint256) {
+    function getSize(uint256 width) public pure returns (uint256) {
         return (width << 1) - numOfPeaks(width);
     }
 
@@ -135,19 +141,27 @@ library MMR {
     /**
      * @dev It returns the hash value of a node for the given position. Note that the index starts from 1
      */
-    function getNode(Tree storage tree, uint256 index) public view returns (bytes32) {
+    function getNode(Tree storage tree, uint256 index)
+        public
+        view
+        returns (bytes32)
+    {
         return tree.hashes[index];
     }
 
     /**
      * @dev It returns a merkle proof for a leaf. Note that the index starts from 1
      */
-    function getMerkleProof(Tree storage tree, uint256 index) public view returns (
-        bytes32 root,
-        uint256 width,
-        bytes32[] memory peakBagging,
-        bytes32[] memory siblings
-    ){
+    function getMerkleProof(Tree storage tree, uint256 index)
+        public
+        view
+        returns (
+            bytes32 root,
+            uint256 width,
+            bytes32[] memory peakBagging,
+            bytes32[] memory siblings
+        )
+    {
         require(index < tree.size, "Out of range");
         require(isLeaf(index), "Not a leaf");
 
@@ -158,7 +172,7 @@ library MMR {
 
         peakBagging = new bytes32[](peaks.length);
         uint256 cursor;
-        for (uint i = 0; i < peaks.length; i++) {
+        for (uint256 i = 0; i < peaks.length; i++) {
             // Collect the hash of all peaks
             peakBagging[i] = tree.hashes[peaks[i]];
             // Find the peak which includes the target index
@@ -190,31 +204,48 @@ library MMR {
         bytes32[] memory itemHashes
     ) public pure returns (bytes32 newRoot) {
         // Check the root equals the peak bagging hash
-        require(root == peakBagging(width, peaks), "Invalid root hash from the peaks");
-        uint tmpWidth = width;
+        require(
+            root == peakBagging(width, peaks),
+            "Invalid root hash from the peaks"
+        );
+        uint256 tmpWidth = width;
         bytes32[255] memory tmpPeakMap = peaksToPeakMap(width, peaks);
-        for (uint i = 0; i < itemHashes.length; i++) {
+        for (uint256 i = 0; i < itemHashes.length; i++) {
             tmpPeakMap = peakUpdate(tmpWidth, tmpPeakMap, itemHashes[i]);
             tmpWidth++;
         }
         return peakBagging(tmpWidth, peakMapToPeaks(tmpWidth, tmpPeakMap));
     }
 
-    function peakBagging(uint256 width, bytes32[] memory peaks) public pure returns (bytes32) {
-        uint size = getSize(width);
-        require(numOfPeaks(width) == peaks.length, "Received invalid number of peaks");
-        return keccak256(abi.encodePacked(size, keccak256(abi.encodePacked(size, peaks))));
+    function peakBagging(uint256 width, bytes32[] memory peaks)
+        public
+        pure
+        returns (bytes32)
+    {
+        uint256 size = getSize(width);
+        require(
+            numOfPeaks(width) == peaks.length,
+            "Received invalid number of peaks"
+        );
+        return
+            keccak256(
+                abi.encodePacked(size, keccak256(abi.encodePacked(size, peaks)))
+            );
     }
 
-    function peaksToPeakMap(uint width, bytes32[] memory peaks) public pure returns (bytes32[255] memory peakMap) {
-        uint bitIndex;
-        uint peakRef;
-        uint count = peaks.length;
-        for(uint height = 1; height <= 255; height++) {
+    function peaksToPeakMap(uint256 width, bytes32[] memory peaks)
+        public
+        pure
+        returns (bytes32[255] memory peakMap)
+    {
+        uint256 bitIndex;
+        uint256 peakRef;
+        uint256 count = peaks.length;
+        for (uint256 height = 1; height <= 255; height++) {
             // Index starts from the right most bit
             bitIndex = 255 - height;
             peakRef = 1 << (height - 1);
-            if((width & peakRef) != 0) {
+            if ((width & peakRef) != 0) {
                 peakMap[bitIndex] = peaks[--count];
             } else {
                 peakMap[bitIndex] = bytes32(0);
@@ -223,12 +254,16 @@ library MMR {
         require(count == 0, "Invalid number of peaks");
     }
 
-    function peakMapToPeaks(uint width, bytes32[255] memory peakMap) public pure returns (bytes32[] memory peaks) {
-        uint arrLength = numOfPeaks(width);
+    function peakMapToPeaks(uint256 width, bytes32[255] memory peakMap)
+        public
+        pure
+        returns (bytes32[] memory peaks)
+    {
+        uint256 arrLength = numOfPeaks(width);
         peaks = new bytes32[](arrLength);
-        uint count = 0;
-        for(uint i = 0; i < 255; i++) {
-            if(peakMap[i] != bytes32(0)) {
+        uint256 count = 0;
+        for (uint256 i = 0; i < 255; i++) {
+            if (peakMap[i] != bytes32(0)) {
                 peaks[count++] = peakMap[i];
             }
         }
@@ -236,25 +271,23 @@ library MMR {
     }
 
     function peakUpdate(
-        uint width,
+        uint256 width,
         bytes32[255] memory prevPeakMap,
         bytes32 itemHash
-    ) public pure returns (
-        bytes32[255] memory nextPeakMap
-    ) {
-        uint newWidth = width + 1;
-        uint cursorIndex = getLeafIndex(newWidth);
-        bytes32 cursorNode = itemHash;//hashLeaf(cursorIndex, itemHash);
-        uint bitIndex;
-        uint peakRef;
+    ) public pure returns (bytes32[255] memory nextPeakMap) {
+        uint256 newWidth = width + 1;
+        uint256 cursorIndex = getLeafIndex(newWidth);
+        bytes32 cursorNode = itemHash; //hashLeaf(cursorIndex, itemHash);
+        uint256 bitIndex;
+        uint256 peakRef;
         bool prevPeakExist;
         bool nextPeakExist;
         bool obtained;
 
-        for(uint height = 1; height <= 255; height++) {
+        for (uint256 height = 1; height <= 255; height++) {
             // Index starts from the right most bit
             bitIndex = 255 - height;
-            if(obtained) {
+            if (obtained) {
                 nextPeakMap[bitIndex] = prevPeakMap[bitIndex];
             } else {
                 peakRef = 1 << (height - 1);
@@ -263,13 +296,13 @@ library MMR {
 
                 // Get new cursor node with hashing the peak and the current cursor
                 cursorIndex++;
-                if(prevPeakExist) {
+                if (prevPeakExist) {
                     cursorNode = hashBranch(prevPeakMap[bitIndex], cursorNode);
                 }
                 // If new peak exists for the bit index
-                if(nextPeakExist) {
+                if (nextPeakExist) {
                     // If prev peak exists for the bit index
-                    if(prevPeakExist) {
+                    if (prevPeakExist) {
                         nextPeakMap[bitIndex] = prevPeakMap[bitIndex];
                     } else {
                         nextPeakMap[bitIndex] = cursorNode;
@@ -295,16 +328,25 @@ library MMR {
         bytes32[] memory peaks,
         bytes32[] memory siblings
     ) public pure returns (bool) {
-        uint size = getSize(width);
+        uint256 size = getSize(width);
         require(size >= index, "Index is out of range");
         // Check the root equals the peak bagging hash
-        require(root == keccak256(abi.encodePacked(size, keccak256(abi.encodePacked(size, peaks)))), "Invalid root hash from the peaks");
+        require(
+            root ==
+                keccak256(
+                    abi.encodePacked(
+                        size,
+                        keccak256(abi.encodePacked(size, peaks))
+                    )
+                ),
+            "Invalid root hash from the peaks"
+        );
 
         // Find the mountain where the target index belongs to
         uint256 cursor;
         bytes32 targetPeak;
         uint256[] memory peakIndexes = getPeakIndexes(width);
-        for (uint i = 0; i < peakIndexes.length; i++) {
+        for (uint256 i = 0; i < peakIndexes.length; i++) {
             if (peakIndexes[i] >= index) {
                 targetPeak = peaks[i];
                 cursor = peakIndexes[i];
@@ -360,7 +402,11 @@ library MMR {
      * @dev It returns the hash a parent node with hash(M | Left child | Right child)
      *      M is the index of the node
      */
-    function hashBranch(bytes32 left, bytes32 right) public pure returns (bytes32) {
+    function hashBranch(bytes32 left, bytes32 right)
+        public
+        pure
+        returns (bytes32)
+    {
         return keccak256(abi.encodePacked(left, right));
     }
 
@@ -368,7 +414,11 @@ library MMR {
      * @dev it returns the hash of a leaf node with hash(M | DATA )
      *      M is the index of the node
      */
-    function hashLeaf(uint256 index, bytes32 dataHash) public pure returns (bytes32) {
+    function hashLeaf(uint256 index, bytes32 dataHash)
+        public
+        pure
+        returns (bytes32)
+    {
         return keccak256(abi.encodePacked(index, dataHash));
     }
 
@@ -409,7 +459,11 @@ library MMR {
     /**
      * @dev It returns the children when it is a parent node
      */
-    function getChildren(uint256 index) public pure returns (uint256 left, uint256 right) {
+    function getChildren(uint256 index)
+        public
+        pure
+        returns (uint256 left, uint256 right)
+    {
         left = index - (uint256(1) << (heightAt(index) - 1));
         right = index - 1;
         require(left != right, "Not a parent");
@@ -419,12 +473,16 @@ library MMR {
      * @dev It returns all peaks of the smallest merkle mountain range tree which includes
      *      the given index(size)
      */
-    function getPeakIndexes(uint256 width) public pure returns (uint256[] memory peakIndexes) {
+    function getPeakIndexes(uint256 width)
+        public
+        pure
+        returns (uint256[] memory peakIndexes)
+    {
         peakIndexes = new uint256[](numOfPeaks(width));
-        uint count;
-        uint size;
-        for(uint i = 255; i > 0; i--) {
-            if(width & (1 << (i - 1)) != 0) {
+        uint256 count;
+        uint256 size;
+        for (uint256 i = 255; i > 0; i--) {
+            if (width & (1 << (i - 1)) != 0) {
                 // peak exists
                 size = size + (1 << i) - 1;
                 peakIndexes[count++] = size;
@@ -433,10 +491,10 @@ library MMR {
         require(count == peakIndexes.length, "Invalid bit calculation");
     }
 
-    function numOfPeaks(uint256 width) public pure returns (uint num) {
+    function numOfPeaks(uint256 width) public pure returns (uint256 num) {
         uint256 bits = width;
-        while(bits > 0) {
-            if(bits % 2 == 1) num++;
+        while (bits > 0) {
+            if (bits % 2 == 1) num++;
             bits = bits >> 1;
         }
         return num;
@@ -448,7 +506,10 @@ library MMR {
      *      it computes hashes recursively downward.
      *      Only appending an item calls this function
      */
-    function _getOrCreateNode(Tree storage tree, uint256 index) private returns (bytes32) {
+    function _getOrCreateNode(Tree storage tree, uint256 index)
+        private
+        returns (bytes32)
+    {
         require(index <= tree.size, "Out of range");
         if (tree.hashes[index] == bytes32(0)) {
             (uint256 leftIndex, uint256 rightIndex) = getChildren(index);
